CMake: Improve usage of assimp DLL/lib path/naming
The name now depends on the local version of Visual Studio rather than a
hardcoded one. It should have a better chance now to work with the next version
of Visual Studio; it hasn’t been tested though.
Using a variable for the DLL path rather than repeating the same value each
time.