Rewrite dependencies handling
Dependencies are now cloned and built in the source folder instead of build
folder, to avoid reclone and rebuild when switching between Debug and Release.
Also force the dependencies to be built in Release mode.
The dependencies handling was rewritten to be more maintainable and do
everything on the first CMake run.

Fixes #25
Fixes #37