Improve resilience of the DLLs copying code
When using the built-in version of CMake in Visual Studio 2017, a few
things are different, like the CMAKE_BINARY_DIR which already includes
the $CONFIG string in the path, and the way we can identify which
version of Visual Studio is being used.
Also factorise the code for copying the DLLs in a single place.