CMake: Find existing previous dependencies
When running CMake on an already existing dependencies folder, the
dependencies were not found as CMAKE_PREFIX_PATH has not been edited.
This would happen if generating the project first as debug in Visual
Studio, and then later switching to release mode, which starts from an
empty cache but an already present dependencies folder.