Trigger a copy of the DLLs for all targets
The DLL was previously only copied for the assignment 1 of EDAF80 which
was fine if the assignments are done in order, but would result in a
missing DLL if a later assignment if done first.
Now trigger the copy for all assignments, and add a check to avoid
copying if the DLL is already present.