Differentiate between minimum and downloaded versions
The minimum version for dependencies is not necessarily the version we
want to use in the labs (as newer versions might contain bug fixes and
such), but we do not want to require higher minimum versions than
needed.
By separating the versions into a minimum set and a set of versions to
use for download, we get the best of both worlds.