Remove Misc
Only very few functions were actually used and the ones that were, were
just single-line wrappers around calls to the C++ standard library.

Take the opportunity to rework the interface of `FPSCamera::Update()` to
take a time duration rather than a `double`, making it explicit in which
units the duration is expected.